// Angular/src/app/models/knjiga.ts
export interface Knjiga{

    id_knjige?: string;

    naziv_knjige?: string; 
   
    opis?: string; 
   
    stanje_na_lageru?: number; 
   
    cena?: number;
   
    zanr?: string; 
   
    ime_autora?: string;
   
    prezime_autora?: string;
    
    slika?: string;

    id_dobavljac?: string; 
}